% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\docType{data}
\name{background_utilities}
\alias{background_utilities}
\title{background_utilities}
\format{
A data frame with 14 rows and 3 variables:
\describe{
  \item{\code{lower}}{double lower age}
  \item{\code{upper}}{double upper age}
  \item{\code{utility}}{double background utility value}
}
}
\usage{
background_utilities
}
\description{
Swedish background utilities from Burstrom and Rehnberg (2006)
}
\details{
http://libris.kb.se/bib/10708053?vw=short
}
\examples{
\dontrun{
if(interactive()){
 background_utilities
 }
}
}
\keyword{datasets}
