% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\docType{data}
\name{cap}
\alias{cap}
\title{DATASET_TITLE}
\format{A data frame with 4 rows and 2 variables:
\describe{
  \item{\code{cohort}}{integer COLUMN_DESCRIPTION}
  \item{\code{pop}}{double COLUMN_DESCRIPTION}
}}
\usage{
cap
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{datasets}
