% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\docType{data}
\name{dre}
\alias{dre}
\title{dre}
\format{
A data frame with 6 rows and 3 variables:
\describe{
  \item{\code{psa_low}}{double Left-hand interval for PSA}
  \item{\code{psa_high}}{double Left-hand interval for PSA}
  \item{\code{dre}}{double total number}
  \item{\code{appa}}{double predicted number with cancer}
  \item{\code{bx}}{double count of the number of biopsies}
  \item{\code{pc}}{double count of the number of cancers}
  \item{\code{sensitivity}}{double sensitivity}
  \item{\code{specificity}}{double specificity}
}
}
\usage{
dre
}
\description{
Sensitivity and specificity for digital rectal examination (DRE)
}
\details{
See https://doi.org/10.1093/jnci/90.23.1817, table 2.
}
\keyword{datasets}
