% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\docType{data}
\name{swedenOpportunisticBiopsyCompliance}
\alias{swedenOpportunisticBiopsyCompliance}
\alias{swedenFormalBiopsyCompliance}
\alias{stockholmTreatment}
\alias{rescreening}
\alias{pop1}
\title{DATASET_TITLE}
\format{A data frame with 15 rows and 3 variables:
\describe{
  \item{\code{psa}}{double COLUMN_DESCRIPTION}
  \item{\code{age}}{double COLUMN_DESCRIPTION}
  \item{\code{compliance}}{double COLUMN_DESCRIPTION}
}}
\usage{
swedenOpportunisticBiopsyCompliance

swedenFormalBiopsyCompliance

stockholmTreatment

rescreening

pop1
}
\description{
DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION
}
\details{
DETAILS

DETAILS

DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{datasets}
