% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\name{ICER.fhcrc}
\alias{ICER.fhcrc}
\title{Calculate an ICER from two fhcrc objects}
\usage{
\method{ICER}{fhcrc}(
  object1,
  object2,
  perspective = c("societal.costs", "healthsector.costs"),
  from = 0,
  ...
)
}
\arguments{
\item{object1}{fhcrc object}

\item{object2}{fhcrc object (reference)}

\item{perspective}{PARAM_DESCRIPTION, Default: c("societal.costs", "healthsector.costs")}

\item{from}{double for the start age (assumes strategies are identical to that age)}

\item{...}{PARAM_DESCRIPTION}
}
\value{
list with perspective, from, ICER.QALE, delta.QALE, delta.costs, and possibly ICER.LE and delta.LE
}
\description{
TBC
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
