% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\name{nn}
\alias{nn}
\title{FUNCTION_TITLE}
\usage{
nn(object1, object2, startAge = 50, stopAge = Inf, ...)
}
\arguments{
\item{object1}{PARAM_DESCRIPTION}

\item{object2}{PARAM_DESCRIPTION}

\item{startAge}{PARAM_DESCRIPTION, Default: 50}

\item{stopAge}{PARAM_DESCRIPTION, Default: Inf}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
