% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\name{plot.fhcrc}
\alias{plot.fhcrc}
\title{FUNCTION_TITLE}
\usage{
\method{plot}{fhcrc}(
  x,
  type = c("incidence.rate", "testing.rate", "biopsy.rate", "metastasis.rate",
    "pc.mortality.rate", "allcause.mortality.rate"),
  plot.type = "l",
  add = FALSE,
  xlab = "Age (years)",
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION, Default: c("incidence.rate", "testing.rate", "biopsy.rate", "metastasis.rate",
"pc.mortality.rate", "allcause.mortality.rate")}

\item{plot.type}{PARAM_DESCRIPTION, Default: 'l'}

\item{add}{PARAM_DESCRIPTION, Default: FALSE}

\item{xlab}{PARAM_DESCRIPTION, Default: 'Age (years)'}

\item{ylab}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
