% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\name{predict.fhcrc}
\alias{predict.fhcrc}
\title{FUNCTION_TITLE}
\usage{
\method{predict}{fhcrc}(
  object,
  scenarios = NULL,
  type = "incidence.rate",
  group = "age",
  group.pt = group,
  age.breaks = NULL,
  year.breaks = NULL,
  cohort.breaks = NULL,
  age.weights = NULL,
  ...
)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{scenarios}{PARAM_DESCRIPTION, Default: NULL}

\item{type}{PARAM_DESCRIPTION, Default: 'incidence.rate'}

\item{group}{PARAM_DESCRIPTION, Default: 'age'}

\item{group.pt}{how to group the person-time, Default: group}

\item{age.breaks}{PARAM_DESCRIPTION, Default: NULL}

\item{year.breaks}{PARAM_DESCRIPTION, Default: NULL}

\item{cohort.breaks}{PARAM_DESCRIPTION, Default: NULL}

\item{age.weights}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
