% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\name{print.fhcrc}
\alias{print.fhcrc}
\title{FUNCTION_TITLE}
\usage{
\method{print}{fhcrc}(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
