% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\docType{data}
\name{sthlm3_mri_arm}
\alias{sthlm3_mri_arm}
\title{Population for the STHLM3-MRI MRI arm}
\format{
A data frame with 30 rows and 2 variables:
\describe{
  \item{\code{cohort}}{birt cohort}
  \item{\code{pop}}{population count}
}
}
\usage{
sthlm3_mri_arm
}
\description{
Birth cohorts for the study, assuming 2019.5 is the middle of the study
}
\examples{
\dontrun{
if(interactive()){
 background_utilities
 }
}
}
\keyword{datasets}
