% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\name{summary.fhcrc}
\alias{summary.fhcrc}
\alias{-.summary.fhcrc}
\alias{*.summary.fhcrc}
\alias{/.summary.fhcrc}
\title{Summarise simulation results}
\usage{
\method{summary}{fhcrc}(object, from = 0, ...)

\method{-}{summary.fhcrc}(object1, object2)

\method{*}{summary.fhcrc}(obj, perpersons)

\method{/}{summary.fhcrc}(obj, denominator)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{from}{Age from which to report}

\item{...}{PARAM_DESCRIPTION}

\item{object1}{The positive reference \code{summary.fhcrc} object.}

\item{object2}{The negative \code{summary.fhcrc} object which is subtracted
from \code{object1}.}

\item{obj}{The \code{summary.fhcrc} object.}

\item{perpersons}{a scalar to multiply the summary results with.}

\item{denominator}{a scalar to divide the summary results with.}
}
\value{
OUTPUT_DESCRIPTION

A \code{summary.fhcrc} object representing the difference between the
    two compared \code{summary.fhcrc}.

A \code{summary.fhcrc} object scaled for a number of persons.

A \code{summary.fhcrc} object scaled by a denominator.
}
\description{
FUNCTION_DESCRIPTION

Calculates the difference between two \code{summary.fhcrc}
    objects.

Calculates the results per a number of persons of a
    \code{summary.fhcrc} object.

Calculates the results divided by a denominator \code{summary.fhcrc} object.
}
\details{
DETAILS

Calculates the difference between two \code{summary.fhcrc} objects,
    e.g. the difference in life expectancy, quality of life etc. The
    exceptions are \code{discountRate.costs} and
    \code{discountRate.effectiveness} which are required two be the same for
    the two scenarios and the original value is returned. If \code{n} has the
    same value the original value is returned if they differ a list with both
    is returned.

N.b. this is unfortunatly not a commutative
    operator. Therefor the first parameter must be the
    \code{summary.fhcrc} object and the second parameter a nummeric
    scalar.

N.b. this is unfortunatly not a commutative
    operator. Therefor the first parameter must be the
    \code{summary.fhcrc} object and the second parameter a nummeric
    scalar.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 scenario1 <- summary(callFhcrc(screen = "screenUptake"))
 scenario2 <-summary(callFhcrc(screen = "noScreening"))
 scenario1 - scenario2
 }
}
\dontrun{
if(interactive()){
 ## Example 1
 scenario1 <- summary(callFhcrc(screen = "screenUptake"))
 scenario1 * 1000 # the summery per thousand persons

 ## Example 2
 scenario2 <-summary(callFhcrc(screen = "noScreening"))
 (scenario1 - scenario2) * 1000 # the difference per thousand persons
 }
}
\dontrun{
if(interactive()){
 }
}
}
