% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\docType{data}
\name{swedenOpportunisticBiopsyCompliance}
\alias{swedenOpportunisticBiopsyCompliance}
\alias{swedenFormalBiopsyCompliance}
\alias{stockholmTreatment}
\alias{rescreening}
\alias{pop1}
\title{DATASET_TITLE}
\format{
A data frame with 15 rows and 3 variables:
\describe{
  \item{\code{psa}}{double COLUMN_DESCRIPTION}
  \item{\code{age}}{double COLUMN_DESCRIPTION}
  \item{\code{compliance}}{double COLUMN_DESCRIPTION}
}

A data frame with 15 rows and 3 variables:
\describe{
  \item{\code{psa}}{double COLUMN_DESCRIPTION}
  \item{\code{age}}{double COLUMN_DESCRIPTION}
  \item{\code{compliance}}{double COLUMN_DESCRIPTION}
}

A data frame with 24 rows and 6 variables:
\describe{
  \item{\code{DxY}}{double COLUMN_DESCRIPTION}
  \item{\code{Age}}{double COLUMN_DESCRIPTION}
  \item{\code{G}}{integer COLUMN_DESCRIPTION}
  \item{\code{CM}}{double COLUMN_DESCRIPTION}
  \item{\code{RP}}{double COLUMN_DESCRIPTION}
  \item{\code{RT}}{double COLUMN_DESCRIPTION}
}

A data frame with 52 rows and 5 variables:
\describe{
  \item{\code{age5}}{double COLUMN_DESCRIPTION}
  \item{\code{total_cat}}{double COLUMN_DESCRIPTION}
  \item{\code{shape}}{double COLUMN_DESCRIPTION}
  \item{\code{cure}}{double COLUMN_DESCRIPTION}
  \item{\code{scale}}{double COLUMN_DESCRIPTION}
}

A data frame with 136 rows and 2 variables:
\describe{
  \item{\code{cohort}}{integer COLUMN_DESCRIPTION}
  \item{\code{pop}}{double COLUMN_DESCRIPTION}
}
}
\usage{
swedenOpportunisticBiopsyCompliance

swedenFormalBiopsyCompliance

stockholmTreatment

rescreening

pop1
}
\description{
DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION
}
\details{
DETAILS

DETAILS

DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{datasets}
