% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\docType{data}
\name{uk_rescreening}
\alias{uk_rescreening}
\title{UK rescreening rates}
\format{A data frame with 42 rows and 5 variables:
\describe{
  \item{\code{age5}}{double COLUMN_DESCRIPTION}
  \item{\code{total_cat}}{double COLUMN_DESCRIPTION}
  \item{\code{shape}}{double COLUMN_DESCRIPTION}
  \item{\code{cure}}{double COLUMN_DESCRIPTION}
  \item{\code{scale}}{double COLUMN_DESCRIPTION}
}}
\usage{
uk_rescreening
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{datasets}
