% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\name{callFhcrc}
\alias{callFhcrc}
\title{Define and run simulation}
\usage{
callFhcrc(
  n = 10,
  screen = "noScreening",
  nLifeHistories = 10,
  seed = 12345,
  panel = FALSE,
  flatPop = FALSE,
  pop = pop1,
  tables = IHE,
  debug = FALSE,
  parms = NULL,
  mc.cores = 1,
  cl = NULL,
  print.timing = TRUE,
  ...
)
}
\arguments{
\item{n}{Integer number of men to simulate. Default: 10}

\item{screen}{String with one of the following simulated screening scenarios:
\describe{
  \item{\code{noScreening}}{no screening test, only diagnosis from symptoms}
  \item{\code{randomScreen50to70}}{TBA}
  \item{\code{twoYearlyScreen50to70}}{two-yearly screening from age 50 to 70}
  \item{\code{fourYearlyScreen50to70}}{four-yearly screening from age 50 to 70}
  \item{\code{screen50}}{one screen at age 50}
  \item{\code{screen60}}{one screen at age 60}
  \item{\code{screen70}}{one screen at age 70}
  \item{\code{screenUptake}}{current testing pattern in Sweden}
  \item{\code{stockholm3_goteborg}}{TBA}
  \item{\code{stockholm3_risk_stratified}}{TBA}
  \item{\code{goteborg}}{risk stratified re-screening 2+4 from age 50 to 70}
  \item{\code{risk_stratified}}{risk stratified re-screening 4+8 from age 50 to 70}
  \item{\code{mixed_screening}}{risk stratified re-screening 2+4 from age 50 to 70 & opportunistic testing for other ages}
  \item{\code{regular_screen}}{TBA}
  \item{\code{single_screen}}{TBA}
  \item{\code{introduced_screening_only}}{TBA}
  \item{\code{introduced_screening_preference}}{TBA}
  \item{\code{introduced_screening}}{TBA}
  \item{\code{stopped_screening}}{TBA}
} . Default: 'noScreening'}

\item{nLifeHistories}{Integer with number of men for all events should be
recorded, Default: 10}

\item{seed}{Integer random number seed, Default: 12345}

\item{panel}{Boolean to use the Stockholm3 biomarker panel test
characteristics, otherwise PSA is used, Default: FALSE}

\item{flatPop}{Boolean to set all birth cohorts of equal size, Default: FALSE}

\item{pop}{Data.frame with two integer columns \code{cohort} with year of the
birth cohorts and \code{pop} with the size of the corresponding birth
cohorts, Default: pop1}

\item{tables}{List of data.frames to update the tables in fhcrcData, Default:
IHE (NB: fhcrcData$ptrx gets changed below to stockholmTreatment if parameters$stockholmTreatment = TRUE -- which is the default! This implies that IHE is the default when parameters$stockholmTreatment = FALSE)}

\item{debug}{Boolean to print debugging, Default: FALSE}

\item{parms}{List to update FhcrcParameters, Default: NULL}

\item{mc.cores}{Integer with the number of cores to use for the computation,
Default: 1}

\item{cl}{a cluster object, created by the \code{parallel} package or by the \code{snow}
package.  If NULL, use \code{mclapply} (i.e. does not use the registered
default cluster).}

\item{print.timing}{Boolean should the required time be printed after the
simulation run, Default: TRUE}

\item{...}{TBA}
}
\value{
A fhcrc object
}
\description{
Function to specify and run the microsimulation. A large number
    of simulated men, \code{n}, will cause the simulation to take longer time
    but will reduce the stochastic variation particularly for rare events.
}
\details{
TBA
}
\examples{
\dontrun{
if(interactive()){
 library(prostata)
 sim1 <- callFhcrc(n=1e6, screen="screenUptake", mc.cores=3)
 }
}
}
\seealso{
\code{\link[parallel]{mclapply}}
}
