% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\docType{data}
\name{cap_study}
\alias{cap_study}
\alias{cap_control}
\title{DATASET_TITLE}
\format{
A data frame with 4 rows and 4 variables:
\describe{
  \item{\code{cohort}}{integer COLUMN_DESCRIPTION}
  \item{\code{pop}}{double COLUMN_DESCRIPTION}
  \item{\code{screened}}{double Number screened at baseline}
  \item{\code{pscreened}}{double Probability of being screened at baseline}
}

A data frame with 4 rows and 4 variables:
\describe{
  \item{\code{cohort}}{integer COLUMN_DESCRIPTION}
  \item{\code{pop}}{double COLUMN_DESCRIPTION}
}
}
\usage{
cap_study

cap_control
}
\description{
DATASET_DESCRIPTION

DATASET_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{datasets}
