% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\docType{data}
\name{fhcrcData}
\alias{fhcrcData}
\title{List tables for the simulation}
\format{A data frame with 12120 rows and 4 variables:
\describe{
  \item{\code{BirthCohort}}{integer COLUMN_DESCRIPTION}
  \item{\code{Age}}{integer COLUMN_DESCRIPTION}
  \item{\code{AnnualMortality}}{double COLUMN_DESCRIPTION}
  \item{\code{Survival}}{double COLUMN_DESCRIPTION}
}
\strong{biopsy_frequency}}
\usage{
fhcrcData
}
\description{
DATASET_DESCRIPTION
\strong{all_cause_mortality}
}
\keyword{datasets}
