% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\docType{data}
\name{FhcrcParameters}
\alias{FhcrcParameters}
\title{Initial values for the FHCRC model}
\format{
A list
\describe{
}
}
\usage{
FhcrcParameters
}
\description{
Initial values for the FHCRC model
}
\details{
Currently, see the R documentation
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{datasets}
