% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\name{ICER.fhcrc}
\alias{ICER.fhcrc}
\title{FUNCTION_TITLE}
\usage{
\method{ICER}{fhcrc}(
  object1,
  object2,
  perspective = c("societal.costs", "healthsector.costs"),
  from = 0,
  ...
)
}
\arguments{
\item{object1}{PARAM_DESCRIPTION}

\item{object2}{PARAM_DESCRIPTION}

\item{perspective}{PARAM_DESCRIPTION, Default: c("societal.costs", "healthsector.costs")}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
