% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostata.R
\docType{package}
\name{prostata-package}
\alias{prostata-package}
\alias{prostata}
\title{prostata}
\section{Introduction}{


The packages provides implementations of discrete event simulation in both R
and C++.
}

\section{Background}{


Prostata is a natural history model of prostate cancer model which extends
the model developed by Ruth Etzioni and colleagues at the
[[http://www.fredhutch.org][Fred Hutchinson Cancer Research Center
(FHCRC)]]. This is a well validated prostate cancer natural history model
developed within the
[[http://cisnet.cancer.gov/prostate/profiles.html][Cancer Intervention and
Surveillance Modeling Network (CISNET)]]. We here provide our extensions of
the model with extended states and finer calibration. The model is provided
as an R package, depending on our
[[https://github.com/mclements/microsimulation][microsimulation]] frame work.
We specifically developed this package for modelling the cost-effectiveness
of prostate cancer screening, where many (e.g. 10^7) men are followed from
birth to death.
}

\section{Running the simulation}{


There are a number of available testing scenarios. They determine testing
frequencies and re-testing intervals over calendar period and ages.
}

\references{
\url{https://github.com/mclements/prostata}

\url{https://github.com/mclements/microsimulation}
}
\seealso{
\code{\link{Rcpp}}
}
\author{
Mark Clements \email{mark.clements@ki.se}
}
