\name{prostata-package}
\alias{prostata-package}
\alias{prostata}
\docType{package}
\title{
  
  Prostata using discrete event simulation. 

 }
 \description{
   
   The packages provides implementations of discrete event simulation in
   both R and C++.

}
\details{
\tabular{ll}{
Package: \tab prostata\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-06-14\cr
License: \tab GPL >= 3.0\cr
}

}
\author{
Mark, Alexandra, Kalle.

Maintainer: Mark Clements <mark.clements@ki.se>
}
% \references{
% ~~ Literature or other references for background information ~~
% }
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% %% ~~ simple examples of the most important functions ~~
% }
